Move MapFilePrimaryExchangeProvider to its own file
